+++ src/renderer/GeneralSettings.tsx
@@ -111,7 +111,7 @@ const GeneralSettings: React.FC<IProps> = (props: IProps) => {
     }
 
     return (
-      <div className="flex flex-col">
+      <div className="flex flex-col w-[500px]">
         <Label htmlFor="storagePath" className="flex items-center">
           {getLocalePhrase(language, Phrase.DiskStorageFolderLabel)}
           <Tooltip
@@ -253,7 +253,7 @@ const GeneralSettings: React.FC<IProps> = (props: IProps) => {
     }
 
     return (
-      <div className="flex flex-col w-1/3 min-w-60 max-w-80">
+      <div className="flex flex-col w-40">
         <Label htmlFor="maxDiskStorage" className="flex items-center">
           {getLocalePhrase(language, Phrase.MaxDiskStorageLabel)}
           <Tooltip
@@ -291,7 +291,7 @@ const GeneralSettings: React.FC<IProps> = (props: IProps) => {
           {getLocalePhrase(language, Phrase.DiskUsageDescription)}
         </Label>
 
-        <div className="flex flex-row items-center justify-start w-1/3 min-w-80 max-w-120 gap-x-2">
+        <div className="flex flex-row items-center justify-start w-80 gap-x-2 py-2">
           <Tooltip
             content={getLocalePhrase(language, Phrase.DiskUsageDescription)}
           >
@@ -309,13 +309,17 @@ const GeneralSettings: React.FC<IProps> = (props: IProps) => {
   return (
     <div className="flex flex-col gap-y-6">
       {getDisabledText()}
-      <div className="flex flex-row gap-x-6">
-        {getStoragePathField()}
-        {getBufferSwitch()}
+      {getStoragePathField()}
+
+      <div className="flex flex-row items-center gap-x-10">
+        {getMaxStorageField()}
+        {getDiskUsageBar()}
       </div>
-      {getBufferPathField()}
-      {getMaxStorageField()}
-      {getDiskUsageBar()}
+
+      <div className="flex flex-row items-center gap-x-10">
+        {getBufferSwitch()}
+        {getBufferPathField()}
+      </div>
     </div>
   );
 };